REF: make pd._testing a directory, separate out asserters
